+++ src/index.ts
@@ -51,9 +51,9 @@ export * from './batch'
 export * from './transaction'
 export * from './fieldValue'
 export * from './fieldPath'
-export * from './onSnapshot/onSnapshot'
+export * from './onSnapshot'
 export * from './operations'
-export * from './queryConstraints'
+export * from './queryClauses'
 export { query } from './refs'
 export * from './equal'
 export type {
