+++ src/queryClauses/endBefore.ts
