+++ src/queryClauses/where.ts
@@ -4,6 +4,7 @@ import {
 	FirelordFirestore,
 	DocumentId,
 	__name__,
+	MetaTypes,
 } from '../types'
 import { documentId } from '../fieldPath'
